+++ libft/ft_strlen.c
@@ -1,4 +1,4 @@
-unsigned long ft_strlen(const char *str)
+size_t  ft_strlen(const char *str)
 {
     int i;
